backend/requirements.txt
pandas
numpy
torch
transformers
sentence-transformers
faiss-cpu
pymongo
python-dotenv
flask
flask-cors
pickle
